Title: 순서에 따른 덧셈, Time: 89ms, Memory: 0MB, Status: Passed - CodeTree

+++ 241009/순서에 따른 덧셈/addition-in-order.cpp
@@ -22,7 +22,7 @@ void input() {
 
 void solve() {
 
-	int result = 0;
+	long long int result = 0;
 
 	for (int i = 0; i < N; i++){
 		int sum = num[i] - i;
